packages/wallet/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
supabase>=2.24.0
web3==6.15.0
eth-abi>=5.0
httpx>=0.27.0
pydantic[email]>=2.11.7
python-dotenv>=1.0.0
resend>=2.0
bcrypt>=4.0
cryptography>=44.0
